v2_services/agent_08_character_art_jury/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
pillow>=10.0.0